feat: Add --dry-run option
Closes #1760.

+++ src/ccache/core/mainoptions.cpp
@@ -113,6 +113,7 @@ Common options:
                                default
     -d, --dir PATH             operate on cache directory PATH instead of the
                                default
+        --dry-run              do not perform any write operations
         --evict-namespace NAMESPACE
                                remove files created in namespace NAMESPACE
         --evict-older-than AGE remove files used less recently than AGE
@@ -309,13 +310,17 @@ print_compression_statistics(const Config& config,
 }
 
 static void
-trim_dir(const std::string& dir,
+trim_dir(core::DryRun dry_run,
+         const std::string& dir,
          const uint64_t trim_max_size,
          const util::SizeUnitPrefixType suffix_type,
          const bool trim_lru_mtime,
          std::optional<std::optional<int8_t>> recompress_level,
          uint32_t threads)
 {
+  ASSERT(dry_run == DryRun::no
+         || recompress_level == std::nullopt); // Verified by caller
+
   std::vector<DirEntry> files;
   uint64_t initial_size = 0;
 
@@ -390,7 +395,7 @@ trim_dir(const std::string& dir,
       if (final_size <= trim_max_size) {
         break;
       }
-      if (util::remove(file.path())) {
+      if (dry_run == DryRun::yes || util::remove(file.path())) {
         ++removed_files;
         final_size -= file.size_on_disk();
       }
@@ -440,6 +445,7 @@ get_usage_text(const std::string_view ccache_name)
 enum : uint8_t {
   CHECKSUM_FILE,
   CONFIG_PATH,
+  DRY_RUN,
   EVICT_NAMESPACE,
   EVICT_OLDER_THAN,
   EXTRACT_RESULT,
@@ -469,6 +475,7 @@ const option long_options[] = {
   {"config-path",             REQUIRED,    nullptr, CONFIG_PATH         },
   {"dir",                     REQUIRED,    nullptr, 'd'                 },
   {"directory",               REQUIRED,    nullptr, 'd'                 }, // compat
+  {"dry-run",                 NO_ARGUMENT, nullptr, DRY_RUN             },
   {"dump-manifest",           REQUIRED,    nullptr, INSPECT             }, // compat
   {"dump-result",             REQUIRED,    nullptr, INSPECT             }, // compat
   {"evict-namespace",         REQUIRED,    nullptr, EVICT_NAMESPACE     },
@@ -517,6 +524,7 @@ process_main_options(int argc, const char* const* argv)
   std::optional<util::SizeUnitPrefixType> trim_suffix_type;
   bool trim_lru_mtime = false;
   std::optional<std::optional<int8_t>> trim_recompress;
+  core::DryRun dry_run = DryRun::no;
 
   std::optional<std::string> evict_namespace;
   std::optional<uint64_t> evict_max_age;
@@ -533,6 +541,9 @@ process_main_options(int argc, const char* const* argv)
     switch (c) {
     case 'd': // --dir
       util::setenv("CCACHE_DIR", arg);
+      break;
+    case DRY_RUN:
+      dry_run = DryRun::yes;
       break;
     case FORMAT:
       if (arg == "tab") {
@@ -599,6 +610,7 @@ process_main_options(int argc, const char* const* argv)
     switch (c) {
     case CONFIG_PATH:
     case 'd': // --dir
+    case DRY_RUN:
     case FORMAT:
     case THREADS:
     case TRIM_MAX_SIZE:
@@ -640,6 +652,11 @@ process_main_options(int argc, const char* const* argv)
     }
 
     case EXTRACT_RESULT: {
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --extract-result\n");
+        return EXIT_FAILURE;
+      }
+
       umask_scope.release(); // Use original umask for files outside cache dir
       const auto cache_entry_data = read_from_path_or_stdin(arg);
       if (!cache_entry_data) {
@@ -693,15 +710,16 @@ process_main_options(int argc, const char* const* argv)
     {
       ProgressBar progress_bar("Cleaning...");
       storage::local::LocalStorage(config).clean_all(
-        [&](double progress) { progress_bar.update(progress); });
-      if (isatty(STDOUT_FILENO)) {
-        PRINT(stdout, "\n");
-      }
+        dry_run, [&](double progress) { progress_bar.update(progress); });
       break;
     }
 
     case 'C': // --clear
     {
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --clear\n");
+        return EXIT_FAILURE;
+      }
       ProgressBar progress_bar("Clearing...");
       storage::local::LocalStorage(config).wipe_all(
         [&](double progress) { progress_bar.update(progress); });
@@ -720,6 +738,10 @@ process_main_options(int argc, const char* const* argv)
       break;
 
     case 'F': { // --max-files
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --max-files\n");
+        return EXIT_FAILURE;
+      }
       auto files = util::value_or_throw<Error>(util::parse_unsigned(arg));
       config.set_value_in_file(
         util::pstr(config.config_path()), "max_files", arg);
@@ -732,6 +754,10 @@ process_main_options(int argc, const char* const* argv)
     }
 
     case 'M': { // --max-size
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --max-size\n");
+        return EXIT_FAILURE;
+      }
       auto [size, suffix_type] =
         util::value_or_throw<Error>(util::parse_size(arg));
       uint64_t max_size = size;
@@ -748,6 +774,10 @@ process_main_options(int argc, const char* const* argv)
     }
 
     case 'o': { // --set-config
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --set-config\n");
+        return EXIT_FAILURE;
+      }
       // Start searching for equal sign at position 1 to improve error message
       // for the -o=K=V case (key "=K" and value "V").
       size_t eq_pos = arg.find('=', 1);
@@ -803,16 +833,26 @@ process_main_options(int argc, const char* const* argv)
     }
 
     case STOP_STORAGE_HELPERS: {
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --stop-storage-helpers\n");
+        return EXIT_FAILURE;
+      }
       storage::Storage storage(config, fs::path(argv[0]).parent_path());
       storage.stop_remote_storage_helpers();
       break;
     }
 
     case TRIM_DIR:
+      if (dry_run == DryRun::yes && trim_recompress) {
+        PRINT(stderr, "--dry-run cannot be used with --trim-recompress\n");
+        return EXIT_FAILURE;
+      }
+
       if (!trim_max_size) {
         throw Error("please specify --trim-max-size when using --trim-dir");
       }
-      trim_dir(arg,
+      trim_dir(dry_run,
+               arg,
                *trim_max_size,
                *trim_suffix_type,
                trim_lru_mtime,
@@ -847,6 +887,10 @@ process_main_options(int argc, const char* const* argv)
 
     case 'X': // --recompress
     {
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --recompress\n");
+        return EXIT_FAILURE;
+      }
       auto wanted_level = parse_compression_level(arg);
 
       ProgressBar progress_bar("Recompressing...");
@@ -858,6 +902,10 @@ process_main_options(int argc, const char* const* argv)
     }
 
     case 'z': // --zero-stats
+      if (dry_run == DryRun::yes) {
+        PRINT(stderr, "--dry-run cannot be used with --zero-stats\n");
+        return EXIT_FAILURE;
+      }
       storage::local::LocalStorage(config).zero_all_statistics();
       PRINT(stdout, "Statistics zeroed\n");
       break;
@@ -874,6 +922,7 @@ process_main_options(int argc, const char* const* argv)
 
     ProgressBar progress_bar("Evicting...");
     storage::local::LocalStorage(config).evict(
+      dry_run,
       [&](double progress) { progress_bar.update(progress); },
       evict_max_age,
       evict_namespace);
